ref RobertLucian/GoPiGo3-dev#2 : feature - fix handling of master

+++ semantic_release.sh
@@ -106,14 +106,20 @@ do_master_release () {
   git log $last_tag HEAD --pretty=format:"%s%n%b" > changelog.txt
   python process_changelog.py $last_tag changelog.txt release.json
 
-  export PYTHONIOENCODING=utf8
-  ID=$(curl -X POST https://api.github.com/repos/$TRAVIS_REPO_SLUG/releases?access_token=$GH_TOKEN --header "Content-Type: application/json" -d @release.json | \
-    python -c "import sys, json; response = json.load(sys.stdin); out = response['id'] if 'id' in response else -1; print(out);")
-  if [[ ID == -1 ]]; then
-    echo "Bad response on publishing release with GitHub API"
-    exit 3
+  if [[ -f release.json ]]; then
+    export PYTHONIOENCODING=utf8
+
+    echo "Creating new release on GitHub on ${TRAVIS_REPO_SLUG}"
+    ID=$(curl -X POST https://api.github.com/repos/$TRAVIS_REPO_SLUG/releases?access_token=$GH_TOKEN --header "Content-Type: application/json" -d @release.json | \
+      python -c "import sys, json; response = json.load(sys.stdin); out = response['id'] if 'id' in response else -1; print(out);")
+    if [[ ID == -1 ]]; then
+      echo "Bad response on publishing release with GitHub API"
+      exit 3
+    else
+      echo $ID > release_id.txt
+    fi
   else
-    echo $ID > release_id.txt
+    echo "Version number not changed, so not creating a new release"
   fi
 }
